Move EquationsOfState into PointwiseFunctions/Hydro
Also combines the EquationsOfState library and the Hydro library
