+++ src/Evolution/Systems/RelativisticEuler/Valencia/PrimitiveFromConservative.hpp
@@ -6,7 +6,7 @@
 #include <cstddef>
 
 #include "DataStructures/Tensor/TypeAliases.hpp"
-#include "PointwiseFunctions/EquationsOfState/EquationOfState.hpp"
+#include "PointwiseFunctions/Hydro/EquationsOfState/EquationOfState.hpp"
 
 /// \cond
 namespace gsl {
