--- src/PointwiseFunctions/EquationsOfState/CMakeLists.txt
@@ -1,19 +1,0 @@
-# Distributed under the MIT License.
-# See LICENSE.txt for details.
-
-set(LIBRARY EquationsOfState)
-
-set(LIBRARY_SOURCES
-  DarkEnergyFluid.cpp
-  IdealFluid.cpp
-  PolytropicFluid.cpp
-  SpecificEnthalpy.cpp
-  )
-
-add_spectre_library(${LIBRARY} ${LIBRARY_SOURCES})
-
-target_link_libraries(
-  ${LIBRARY}
-  INTERFACE DataStructures
-  INTERFACE ErrorHandling
-  )
