+++ tests/Unit/PointwiseFunctions/Hydro/CMakeLists.txt
@@ -4,6 +4,10 @@
 set(LIBRARY "Test_Hydro")
 
 set(LIBRARY_SOURCES
+  EquationsOfState/Test_DarkEnergyFluid.cpp
+  EquationsOfState/Test_IdealFluid.cpp
+  EquationsOfState/Test_PolytropicFluid.cpp
+  EquationsOfState/Test_SpecificEnthalpy.cpp
   Test_LorentzFactor.cpp
   Test_Tags.cpp
   )
